Failed project
There is a issue with the project not beeing compatible with the browser because of manifest v2 tried to convert it to manifest v3 but ran into alot of issues.

Tried searching for solutions for a few hours but ran into alot of bugs and could not finish the project. putting it on the side for the moment but maybe gonna come back to take a fresh look at it later this week.